[feat/#18] 채팅방 생성 API data 모듈 구성
