--- core/data/src/commonMain/kotlin/com/nexters/emotia/core/data/chat/ChattingRemoteDataSource.kt
@@ -1,7 +1,0 @@
-package com.nexters.emotia.core.data.chat
-
-import com.nexters.emotia.network.FakeResponse
-
-interface ChattingRemoteDataSource {
-    suspend fun createRoom(): FakeResponse
-}
