--- core/data/src/commonMain/kotlin/com/nexters/emotia/core/data/chat/ChattingRemoteDataSourceImpl.kt
@@ -1,13 +1,0 @@
-package com.nexters.emotia.core.data.chat
-
-import com.nexters.emotia.network.ApiService
-import com.nexters.emotia.network.FakeResponse
-
-class ChattingRemoteDataSourceImpl(
-    private val apiService: ApiService
-) : ChattingRemoteDataSource {
-    
-    override suspend fun createRoom(): FakeResponse {
-        return apiService.getGeolocationInfo("24.48.0.1")
-    }
-}
