+++ core/data/src/commonMain/kotlin/com/nexters/emotia/core/data/chat/datasource/ChattingRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package com.nexters.emotia.core.data.chat.datasource
+
+import com.nexters.emotia.network.dto.response.CreateRoomResponse
+
+interface ChattingRemoteDataSource {
+    suspend fun createRoom(username: String): CreateRoomResponse
+}
