--- core/network/src/commonMain/kotlin/com/nexters/emotia/network/ApiService.kt
@@ -1,9 +1,0 @@
-package com.nexters.emotia.network
-
-class ApiService(
-    private val network: EmotiaNetwork,
-) {
-    suspend fun getGeolocationInfo(ipAddress: String): FakeResponse {
-        return network.get("json/$ipAddress")
-    }
-}
